fix: move validate_schedule to policy.rs so it compiles without db feature
builder.rs is always compiled (no db feature gate) but was calling
crate::scheduler::validate_schedule which is db-gated. Moving the
implementation to policy.rs (always compiled, croner is unconditional)
and re-exporting from scheduler.rs fixes the --no-default-features build.
