+++ autumn-harvest/src/lib.rs
@@ -110,6 +110,7 @@ pub use execution::{
 pub use executor::{WorkflowOutcome, run_workflow};
 pub use history_export::export_mermaid_sequence;
 pub use info::{ActivityHandlerFn, ActivityInfo, DagInfo, WorkflowHandlerFn, WorkflowInfo};
+pub use policy::validate_schedule;
 pub use policy::{RetryPolicy, Schedule, TaskStatus, TriggerRule, WorkflowSchedule};
 pub use pool::{HarvestPoolConfig, compute_pool_sizes};
 pub use query::QueryRegistry;
@@ -121,7 +122,7 @@ pub use saga::Saga;
 #[cfg(feature = "db")]
 pub use scheduler::{
     DagCatalog, RegisteredDag, SchedulerMonitor, SchedulerRuntime, compile_dag_catalog,
-    register_schedules, register_workflow_schedules, tick_once, trigger_dag, validate_schedule,
+    register_schedules, register_workflow_schedules, tick_once, trigger_dag,
 };
 pub use shard::ShardRouter;
 #[cfg(feature = "db")]
